chore: cleanup of scripts and refactor team aura
