+++ Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:2ad0da151ae5faff47e7c0092c91ea0e81e3c2192ecb2b82ae7d351646eb345c
-size 1413
+oid sha256:224b059d7bac00aebbd173295acbe17b18ce94bfc00fe4e229b2620545517bb6
+size 1447
