+++ Assets/Prefabs/team_aura.prefab
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:132df544118daa9295b4ff6d44741f9784d79ad2ccf75609c342ca5bca1dd3a0
-size 116496
+oid sha256:0a11504e6485f0aba2a4c3c3d6378f7173a299b6ca93ae38e34167d3e8bb59e4
+size 117138
