+++ Assets/Resources/Encanto.prefab
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:66a02c8072522586b5441392553083c72dee9894a2a1718d791900525e7c0ff6
-size 183778
+oid sha256:056298f3efe4b88df4b5c49ded203692e5c8e8541cccd088dc1e76a317e9e786
+size 297876
