+++ Assets/Resources/Mangkukulam.prefab
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:e5762c55bca8c207da387a845143af81dd3dc5bad91a8e72db389521107a8958
-size 305361
+oid sha256:45f8d6e5e1e9d5699d8ef4d6e9cb1fcd5f5a91a93ae5f869a89a0cb2a15f7b22
+size 419463
