+++ Assets/Resources/Nuno.prefab
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:949f80de520d96e52e73c672d1604ac340dfed54f82555315e9a70c9766bc142
-size 277674
+oid sha256:39f9887b56c6088cfe81e6e74ef59221176372ef62acecb81dd235c31c2ffae9
+size 391776
