+++ Assets/Resources/Tikbalang.prefab
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0180ddf36de5330c3abd45c144a68eddfa714e94f82788b0b1555f53f1bc9591
-size 294980
+oid sha256:7a234488e87d0b7864ece50da927d2e72f6c90aa47137446c0b35759e59d08ba
+size 409079
