+++ Assets/Scripts/ScoreManager.cs
