+++ Assets/Scripts/ScoreManager.cs.meta
