+++ Assets/Scripts/TeamAura.cs.meta
