+++ Assets/Scripts/VictorySingleton.cs
