+++ ProjectSettings/ProjectSettings.asset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:88d589507963822100e83860f4f855cfb7cac715e659ee29e9a620b7a89f0a70
-size 21632
+oid sha256:058b0ec028e461a6dbce445d43b765587d0d36680d446ba7ac2c65ffcaeda646
+size 21701
